+++ src/main/java/com/mobiquity/jsonplaceholder/config/Configuration.java
@@ -1,0 +1,16 @@
+package com.mobiquity.jsonplaceholder.config;
+
+import org.aeonbits.owner.Config;
+import org.aeonbits.owner.Config.LoadPolicy;
+import org.aeonbits.owner.Config.LoadType;
+
+@LoadPolicy(LoadType.MERGE)
+@Config.Sources({
+        "system:properties",
+        "classpath:configuration.properties"})
+public interface Configuration extends Config {
+
+    @Key("api.base.uri")
+    String baseURI();
+
+}