+++ src/test/java/com/mobiquity/jsonplaceholder/comments/CommentsContractTest.java
@@ -2,6 +2,8 @@ package com.mobiquity.jsonplaceholder.comments;
 
 import com.mobiquity.jsonplaceholder.BaseAPI;
 import org.testng.annotations.Test;
+
+import static com.mobiquity.jsonplaceholder.config.RelativeURI.GET_ALL_COMMENTS;
 import static io.restassured.RestAssured.when;
 import static io.restassured.module.jsv.JsonSchemaValidator.matchesJsonSchemaInClasspath;
 
@@ -12,7 +14,7 @@ public class CommentsContractTest extends BaseAPI {
     public void getPostCommentsSimulation(){
 
             when().
-                    get("https://jsonplaceholder.typicode.com/comments").
+                    get(GET_ALL_COMMENTS).
                     then().
                     body(matchesJsonSchemaInClasspath("schemas/posts_comments_simulation.json"));