+++ src/test/java/com/mobiquity/jsonplaceholder/posts/PostsContractTest.java
@@ -3,6 +3,9 @@ package com.mobiquity.jsonplaceholder.posts;
 import com.mobiquity.jsonplaceholder.BaseAPI;
 import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
+
+import static com.mobiquity.jsonplaceholder.config.RelativeURI.GET_POST_COMMENTS;
+import static com.mobiquity.jsonplaceholder.config.RelativeURI.GET_POST_DETAILS;
 import static io.restassured.RestAssured.when;
 import static io.restassured.module.jsv.JsonSchemaValidator.matchesJsonSchemaInClasspath;
 
@@ -11,7 +14,7 @@ public class PostsContractTest extends BaseAPI {
     @Test
     public void getPostsSimulation(){
         when().
-                get("https://jsonplaceholder.typicode.com/posts").
+                get(GET_POST_DETAILS).
                 then().
                 body(matchesJsonSchemaInClasspath("schemas/posts_simulation.json"));
     }
@@ -19,7 +22,7 @@ public class PostsContractTest extends BaseAPI {
     @Test
     public void getPostCommentsSimulation(){
         when().
-                get("https://jsonplaceholder.typicode.com/posts/99/comments").
+                get(GET_POST_COMMENTS.replace("{PostID}","99")).
                 then().
                 body(matchesJsonSchemaInClasspath("schemas/posts_comments_simulation.json"));
     }