+++ src/test/java/com/mobiquity/jsonplaceholder/users/UsersContractTest.java
@@ -2,6 +2,9 @@ package com.mobiquity.jsonplaceholder.users;
 
 import com.mobiquity.jsonplaceholder.BaseAPI;
 import org.testng.annotations.Test;
+
+import static com.mobiquity.jsonplaceholder.config.RelativeURI.GET_USER_ALBUMS;
+import static com.mobiquity.jsonplaceholder.config.RelativeURI.GET_USER_TODOS;
 import static io.restassured.RestAssured.when;
 import static io.restassured.module.jsv.JsonSchemaValidator.matchesJsonSchemaInClasspath;
 
@@ -11,7 +14,7 @@ public class UsersContractTest extends BaseAPI{
     @Test
     public void getUserAlbumsModel(){
         when().
-                get("https://jsonplaceholder.typicode.com/users/1/albums").
+                get(GET_USER_ALBUMS.replace("{UserId}","1")).
                 then().
                 body(matchesJsonSchemaInClasspath("schemas/user_albums_simulation.json"));
     }
@@ -19,7 +22,7 @@ public class UsersContractTest extends BaseAPI{
     @Test
     public void getUserTodosModel(){
         when().
-                get("https://jsonplaceholder.typicode.com/users/1/todos").
+                get(GET_USER_TODOS.replace("{UserId}","1")).
                 then().
                 body(matchesJsonSchemaInClasspath("schemas/user_todos_simulation.json"));
     }