+++ CHANGELOG.md
@@ -10,5 +10,13 @@
  - Added schemas
  - added gitignore files
  
+## [1.0.1] 30-04-2022
 
-  
+### Changed
+   - Added libraries
+       - owner (for Configuration Autobind)
+   - Added classes
+        - Configuration Interface 
+        - Relative URI
+        - BaseAPI Updated
+